Initial Commit
This function solves for differences between y-values (ex: first diff, second diff, etc) as far as the given data allows and displays the degree of a function. This works, given that the x values in the table are evenly spaced. I am considering adding more features but this was the originally intended goal for this project